fixed schema create order. prevent NPE when trying to delete a witness with no xslt

+++ juxta-ws/src/main/java/org/juxtasoftware/dao/impl/JuxtaXsltDaoImpl.java
@@ -26,7 +26,9 @@ public class JuxtaXsltDaoImpl extends JuxtaDaoImpl<JuxtaXslt> implements JuxtaXs
 
     @Override
     public void delete(JuxtaXslt juxtaXslt) {
-        this.jt.update("delete from " + this.tableName + " where id = ?", juxtaXslt.getId());
+        if ( juxtaXslt != null ) {
+            this.jt.update("delete from " + this.tableName + " where id = ?", juxtaXslt.getId());
+        }
     }
 
     @Override
